Convert HA-STATE-LABEL-BADGE to lit
